renamed packages, class and logged exception
