+++ src/main/java/it/eng/idsa/businesslogic/processor/receiver/ReceiverUsageControlProcessor.java
@@ -30,11 +30,11 @@ import com.google.gson.stream.JsonWriter;
 import de.fraunhofer.iais.eis.ArtifactRequestMessage;
 import de.fraunhofer.iais.eis.ArtifactResponseMessage;
 import de.fraunhofer.iais.eis.Message;
-import it.eng.idsa.businesslogic.camel_interceptor_ucapp.model.Meta;
-import it.eng.idsa.businesslogic.camel_interceptor_ucapp.model.TargetArtifact;
-import it.eng.idsa.businesslogic.camel_interceptor_ucapp.model.UsageControlObject;
 import it.eng.idsa.businesslogic.service.MultipartMessageService;
 import it.eng.idsa.businesslogic.service.RejectionMessageService;
+import it.eng.idsa.businesslogic.usagecontrol.model.Meta;
+import it.eng.idsa.businesslogic.usagecontrol.model.TargetArtifact;
+import it.eng.idsa.businesslogic.usagecontrol.model.UsageControlObject;
 import it.eng.idsa.businesslogic.util.RejectionMessageType;
 import it.eng.idsa.multipart.builder.MultipartMessageBuilder;
 import it.eng.idsa.multipart.domain.MultipartMessage;
