+++ src/main/java/it/eng/idsa/businesslogic/usagecontrol/model/IdsMsgTarget.java
@@ -1,4 +1,4 @@
-package it.eng.idsa.businesslogic.camel_interceptor_ucapp.model;
+package it.eng.idsa.businesslogic.usagecontrol.model;
 
 public class IdsMsgTarget {
 	String name;
