+++ src/main/java/it/eng/idsa/businesslogic/usagecontrol/model/IdsUseObject.java
@@ -1,4 +1,4 @@
-package it.eng.idsa.businesslogic.camel_interceptor_ucapp.model;
+package it.eng.idsa.businesslogic.usagecontrol.model;
 
 public class IdsUseObject {
