+++ src/main/java/it/eng/idsa/businesslogic/usagecontrol/model/UsageControlObject.java
@@ -1,4 +1,4 @@
-package it.eng.idsa.businesslogic.camel_interceptor_ucapp.model;
+package it.eng.idsa.businesslogic.usagecontrol.model;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.apache.commons.lang3.builder.HashCodeBuilder;
