+++ src/test/java/it/eng/idsa/businesslogic/processor/sender/SenderUsageControlProcessorTest.java
@@ -23,11 +23,11 @@ import org.springframework.test.util.ReflectionTestUtils;
 import com.google.gson.internal.LinkedTreeMap;
 
 import de.fraunhofer.iais.eis.Message;
-import it.eng.idsa.businesslogic.camel_interceptor_ucapp.model.IdsUseObject;
-import it.eng.idsa.businesslogic.camel_interceptor_ucapp.service.UcService;
 import it.eng.idsa.businesslogic.processor.exception.ExceptionForProcessor;
 import it.eng.idsa.businesslogic.service.RejectionMessageService;
 import it.eng.idsa.businesslogic.service.impl.RejectionMessageServiceImpl;
+import it.eng.idsa.businesslogic.usagecontrol.model.IdsUseObject;
+import it.eng.idsa.businesslogic.usagecontrol.service.UcService;
 import it.eng.idsa.businesslogic.util.HeaderCleaner;
 import it.eng.idsa.businesslogic.util.MessagePart;
 import it.eng.idsa.businesslogic.util.RejectionMessageType;
